+++ src/field.cpp
@@ -145,9 +145,10 @@ void field_debug(bool *isOpen)
 		return;
 	}
 
-	ImGui::Text("Current field ID: %d", *common_externals.current_field_id);
+	if (!ff8) ImGui::Text("Game Moment: %u", *ff7_externals.field_game_moment);
+	ImGui::Text("Current field ID: %u", *common_externals.current_field_id);
 	if (!ff8) ImGui::Text("Current triangle ID: %d", *ff7_externals.current_triangle_id);
-	ImGui::Text("Previous field ID: %d", *common_externals.previous_field_id);
+	ImGui::Text("Previous field ID: %u", *common_externals.previous_field_id);
 	ImGui::Separator();
 
 	// Inputs for changing field map